+++ exercises/structs/structs1.rs
@@ -5,16 +5,18 @@
 // Execute `rustlings hint structs1` or use the `hint` watch subcommand for a
 // hint.
 
+// 定义经典C风格结构体：包含命名字段
 struct ColorClassicStruct {
-    red: i32,
-    green: i32,
-    blue: i32,
+    red: u8,    // RGB颜色的红色通道（0-255）
+    green: u8,  // 绿色通道
+    blue: u8,   // 蓝色通道
 }
 
-struct ColorTupleStruct(i32, i32, i32);
+// 定义元组结构体：字段没有名称，只有类型
+struct ColorTupleStruct(u8, u8, u8);
 
 #[derive(Debug)]
-struct UnitLikeStruct;
+struct UnitLikeStruct;  // 单元结构体：没有任何字段
 
 #[cfg(test)]
 mod tests {
@@ -22,6 +24,7 @@ mod tests {
 
     #[test]
     fn classic_c_structs() {
+        // 实例化经典C风格结构体
         let green = ColorClassicStruct {
             red: 0,
             green: 255,
@@ -35,6 +38,7 @@ mod tests {
 
     #[test]
     fn tuple_structs() {
+        // 实例化元组结构体（按位置赋值）
         let green = ColorTupleStruct(0, 255, 0);
 
         assert_eq!(green.0, 0);
@@ -44,6 +48,7 @@ mod tests {
 
     #[test]
     fn unit_structs() {
+        // 实例化单元结构体（无字段，直接写结构体名）
         let unit_like_struct = UnitLikeStruct;
         let message = format!("{:?}s are fun!", unit_like_struct);
